updated match_outputs and started gather_energies

+++ src/match_outputs.py
@@ -9,8 +9,20 @@ for i in directories:
     os.chdir(i)
     out_files = glob.glob("*.out*")
     out_completion = glob.glob("mex_o.*")
-    for i in range(len(out_completion)-len(out_files)):
-        del_o = out_completion.pop()
-        subprocess.call("echo " + del_o, shell=True)
-        subprocess.call("rm " + del_o, shell=True)
+    len_file = len(out_files)
+    len_complete = len(out_completion)
+    #print("len_complete", len(out_completion), "len_outfiles", len(out_files))
+    if len_complete > len_file:
+        for i in range(len_complete-len_file):
+            del_o = out_completion.pop()
+            #subprocess.call("echo " + del_o, shell=True)
+            print('rm %s' % del_o)
+            subprocess.call("rm " + del_o, shell=True)
+    elif len_complete < len_file:
+        for i in range((len_file - len_complete)):
+            create_o = "mex_o.o%s0000" % str(i+15)
+            print('touch %s' % create_o)
+            #subprocess.call("echo " + create_o, shell=True)
+            subprocess.call("touch " + create_o, shell=True)
+
     os.chdir("..")
